requirements.txt
streamlit
tweepy
pandas
matplotlib
seaborn
plotly
textblob
python-dotenv
streamlit-autorefresh